added LFs

+++ tutorials/organic_synthesis_figures/parse.py
@@ -24,7 +24,7 @@ from fonduer import HTMLPreprocessor, OmniParser
 docs_path = os.environ['FONDUERHOME'] + 'tutorials/organic_synthesis_figures/data/html/'
 pdf_path = os.environ['FONDUERHOME'] + 'tutorials/organic_synthesis_figures/data/pdf/'
 
-max_docs = float(10)
+max_docs = float(3)
 doc_preprocessor = HTMLPreprocessor(docs_path, max_docs=max_docs)
 corpus_parser = OmniParser(structural=True, lingual=True, visual=True, pdf_path=pdf_path,
 #                           flatten=['sup', 'sub', 'small'],
